create models module #55
