+++ esaapi/models/star.go
@@ -1,0 +1,9 @@
+package models
+
+import "time"
+
+type Stargazer struct {
+	CreatedAt *time.Time `json:"created_at"`
+	Body      string
+	User      User
+}
