+++ esaapi/models/team.go
@@ -1,0 +1,9 @@
+package models
+
+type Team struct {
+	Name        string `json:"name"`
+	Privacy     string `json:"privacy"`
+	Description string `json:"description"`
+	Icon        string `json:"icon"`
+	URL         string `json:"url"`
+}
