+++ esaapi/teams/types/teams_response.go
@@ -3,11 +3,12 @@ package types
 import (
 	"net/http"
 
+	"github.com/michimani/go-esa/esaapi/models"
 	"github.com/michimani/go-esa/gesa"
 )
 
 type TeamsGetResponse struct {
-	Teams      []Team           `json:"teams"`
+	Teams      []models.Team    `json:"teams"`
 	PrevPage   *gesa.PageNumber `json:"prev_page,omitempty"`
 	NextPage   *gesa.PageNumber `json:"next_page,omitempty"`
 	TotalCount int              `json:"total_count"`
@@ -18,14 +19,6 @@ type TeamsGetResponse struct {
 	RateLimitInfo *gesa.RateLimitInformation `json:"-"`
 }
 
-type Team struct {
-	Name        string `json:"name"`
-	Privacy     string `json:"privacy"`
-	Description string `json:"description"`
-	Icon        string `json:"icon"`
-	URL         string `json:"url"`
-}
-
 func (r *TeamsGetResponse) SetRateLimitInfo(h http.Header) {
 	if rri, err := gesa.GetRateLimitInformation(h); err == nil {
 		r.RateLimitInfo = rri
@@ -33,7 +26,7 @@ func (r *TeamsGetResponse) SetRateLimitInfo(h http.Header) {
 }
 
 type TeamsTeamNameGetResponse struct {
-	Team
+	models.Team
 
 	RateLimitInfo *gesa.RateLimitInformation `json:"-"`
 }
